test(DataGrid) STAR-248 fixing and cleaning tests
