+++ packages/DataGrid/tests/DataGrid.cypress.js
@@ -1,24 +1,14 @@
-describe("DataGrid", () => {
+const keyEvent = {
+  enter: { keyCode: 13, which: 13, force: true },
+  up: { keyCode: 38, which: 38, force: true },
+  down: { keyCode: 40, which: 40, force: true },
+};
+
+describe("<DataGrid />", () => {
   it("Should have custom css", () => {
     cy.visitStorybook("data-table-datagrid-backyard-tests--custom-css")
       .get('[data-column-index="1"]')
       .should("have.css", "background-color", "rgb(204, 229, 253)");
-  });
-
-  it("Should autofocus the last selection", () => {
-    cy.visitStorybook("data-table-datagrid-examples--autofocus")
-      /* Selecting and click the item does not appear to work in cypress version 3.3.1. It does however work when selected and clicked twice. 
-    When updating cypress the repetition can and should be removed to avoid testing fails. */
-      // <---from here--->
-      .wait(100)
-      .get('[data-row-index="0"]')
-      .click();
-    // <---till here--->
-    cy.get('[data-row-index="0"]').click();
-    cy.get("button")
-      .click()
-      .focused()
-      .contains("Josef Bican ‡");
   });
 
   it("Should run a callback function", () => {
@@ -57,21 +47,110 @@ describe("DataGrid", () => {
       .should("not.be.visible");
   });
 
-  it("Should show expanded content", () => {
-    cy.visitStorybook("data-table-datagrid-examples--collapsible")
-      /* Selecting and click the item does not appear to work in cypress version 3.3.1. It does however work when selected and clicked twice. 
-    When updating cypress the repetition can and should be removed to avoid testing fails. */
-      // <---from here--->
-      .getByRole("grid")
-      .get('[data-row-index="0"]')
+  it("should render StickyColumn", () => {
+    cy.visitStorybook("data-table-datagrid-examples--lazy");
+    cy.get('[data-column-index="2"]')
       .eq(0)
-      .click();
-    // <---till here--->
-    cy.getByRole("grid")
-      .get('[data-row-index="0"]')
+      .then(e => {
+        const coords = e[0].getBoundingClientRect();
+
+        cy.getAllByRole("rowgroup")
+          .last()
+          .scrollTo(100, 0);
+
+        cy.get('[data-column-index="2"]')
+          .eq(0)
+          .then(e => {
+            const checkcoords = e[0].getBoundingClientRect();
+            expect(checkcoords).to.deep.equal(coords);
+          });
+      });
+
+    cy.getAllByRole("rowgroup")
+      .last()
+      .scrollTo(-100, 0);
+
+    cy.get('[data-column-index="5"]')
       .eq(0)
+      .then(e => {
+        const coords = e[0].getBoundingClientRect();
+
+        cy.getAllByRole("rowgroup")
+          .last()
+          .scrollTo(200, 0);
+
+        cy.get('[data-column-index="5"]')
+          .eq(0)
+          .then(e => {
+            const checkcoords = e[0].getBoundingClientRect();
+            expect(checkcoords).to.not.deep.equal(coords);
+          });
+      });
+  });
+
+  it("should render ColumnIndicator", () => {
+    cy.visitStorybook("data-table-datagrid-examples--lazy");
+    cy.get('[data-row-index="0"]')
+      .eq(0)
+      .children()
+      .children()
+      .should($div => {
+        const className = $div[0].className;
+        expect(className).to.match(/Checkbox/);
+      });
+  });
+
+  it("should render ColumnExpand", () => {
+    cy.visitStorybook("data-table-datagrid-examples--lazy");
+    cy.get('[data-column-index="1"]')
+      .eq(0)
+      .children()
+      .first()
+      .should("have.css", "opacity", "0");
+
+    cy.get('[data-column-index="1"]')
+      .eq(0)
+      .children()
+      .first()
+      .trigger("mouseover")
+      .should("have.css", "opacity", "1");
+  });
+
+  it("should render load more button when you scroll to the bottom", () => {
+    cy.visitStorybook("data-table-datagrid-examples--lazy");
+    cy.getAllByRole("rowgroup")
+      .last()
+      .scrollTo("bottom")
+      .should("not.contain", "Albion")
+      .should("contain", "Abyss");
+
+    cy.get('[data-pka-anchor="button"')
+      .should("be.visible")
       .click()
-      .getAllByText(/narratives/i)
-      .should("be.visible");
+      .wait(200)
+      .getAllByRole("rowgroup")
+      .last()
+      .scrollTo("bottom")
+      .contains("Albion");
+  });
+
+  it("should navigate DataGrid-collapsible with enter key and up, down arrow keys", () => {
+    cy.visitStorybook("data-table-datagrid-examples--collapsible");
+
+    cy.getByText("Audit Planning")
+      .trigger("mouseup")
+      .getAllByRole("grid")
+      .should("contain", "Narratives");
+
+    cy.getByText("Audit Planning")
+      .trigger("mouseup")
+      .trigger("keydown", keyEvent.down)
+      .trigger("keyup", keyEvent.enter)
+      .trigger("keydown", keyEvent.down)
+      .trigger("keydown", keyEvent.down)
+      .get(".grid--is-active [role='gridcell']")
+      .should($div => {
+        expect($div.text()).to.match(/Card Issuance and Management 2/i);
+      });
   });
 });
